Move Log to oplog directory, refactor references
