+++ src/oplog/index.js
@@ -1,0 +1,4 @@
+export { default as Log, DefaultAccessController } from './log.js'
+export { default as Entry } from './entry.js'
+export { default as Clock } from './lamport-clock.js'
+export { default as Sorting } from './sorting.js'
