+++ src/oplog/lamport-clock.js
