+++ src/oplog/log.js
@@ -2,13 +2,13 @@ import LRU from 'lru'
 import Entry from './entry.js'
 import Clock from './lamport-clock.js'
 import Heads from './heads.js'
-import Sorting from './sorting/log.js'
-import MemoryStorage from './storage/memory.js'
+import Sorting from './sorting.js'
+import MemoryStorage from '../storage/memory.js'
 // import LRUStorage from './storage/lru.js'
 // import LevelStorage from './storage/level.js'
 // import IPFSBlockStorage from './storage/ipfs-block.js'
 // import ComposedStorage from './storage/composed.js'
-import { isDefined } from './utils/index.js'
+import { isDefined } from '../utils/index.js'
 
 const { LastWriteWins, NoZeroes } = Sorting
