--- src/sorting/index.js
@@ -1,1 +1,0 @@
-export { LogSorting } from './src/log.js'
