+++ test/keyvalue.spec.js
@@ -1,7 +1,6 @@
 import { deepStrictEqual, strictEqual } from 'assert'
 import rimraf from 'rimraf'
-import Log from '../src/log.js'
-import Entry from '../src/entry.js'
+import { Log, Entry } from '../src/oplog/index.js'
 import IdentityProvider from '../src/identities/index.js'
 import KeyStore from '../src/key-store.js'
 import { KeyValue, KeyValuePersisted, Database } from '../src/db/index.js'
