+++ test/oplog/entry.test.js
@@ -1,7 +1,7 @@
 import { strictEqual, deepStrictEqual } from 'assert'
 import rimraf from 'rimraf'
 import { copy } from 'fs-extra'
-import Entry from '../../src/entry.js'
+import { Entry } from '../../src/oplog/index.js'
 import IdentityProvider from '../../src/identities/index.js'
 import KeyStore from '../../src/key-store.js'
 import { config, testAPIs, startIpfs, stopIpfs } from 'orbit-db-test-utils'
