+++ test/oplog/join-concurrent.test.js
@@ -1,7 +1,7 @@
 import { strictEqual, deepStrictEqual } from 'assert'
 import rimraf from 'rimraf'
 import { copy } from 'fs-extra'
-import Log from '../../src/log.js'
+import { Log } from '../../src/oplog/index.js'
 import IdentityProvider from '../../src/identities/index.js'
 import KeyStore from '../../src/key-store.js'
 import MemoryStorage from '../../src/storage/memory.js'
