+++ test/oplog/join.test.js
@@ -1,7 +1,6 @@
 import { strictEqual, notStrictEqual, deepStrictEqual } from 'assert'
 import rimraf from 'rimraf'
-import Clock from '../../src/lamport-clock.js'
-import Log from '../../src/log.js'
+import { Log, Clock } from '../../src/oplog/index.js'
 import IdentityProvider from '../../src/identities/index.js'
 import KeyStore from '../../src/key-store.js'
 import MemoryStorage from '../../src/storage/memory.js'
