+++ test/oplog/load.test.js
@@ -1,10 +1,8 @@
 import { strictEqual, deepStrictEqual, notStrictEqual, throws } from 'assert'
 import rimraf from 'rimraf'
 import { copy } from 'fs-extra'
-import Sorting from '../../src/sorting/log.js'
+import { Log, Entry, Sorting } from '../../src/oplog/index.js'
 import bigLogString from '../fixtures/big-log.fixture.js'
-import Entry from '../../src/entry.js'
-import Log from '../../src/log.js'
 import IdentityProvider from '../../src/identities/index.js'
 import KeyStore from '../../src/key-store.js'
 import LogCreator from './utils/log-creator.js'
