+++ test/oplog/log.spec.js
@@ -1,7 +1,6 @@
 import { notStrictEqual, deepStrictEqual, strictEqual } from 'assert'
 import rimraf from 'rimraf'
-import Entry from '../../src/entry.js'
-import Log from '../../src/log.js'
+import { Log, Entry } from '../../src/oplog/index.js'
 import IdentityProvider from '../../src/identities/index.js'
 import KeyStore from '../../src/key-store.js'
 import { copy } from 'fs-extra'
